+++ src/ui_linedef.cc
@@ -843,7 +843,7 @@ void UI_LineBox::CalcLength()
 
 	int n = obj;
 
-	float len_f = LineDefs[n]->CalcLength();
+	float len_f = static_cast<float>(LineDefs[n]->CalcLength());
 
 	char buffer[128];
 	snprintf(buffer, sizeof(buffer), "%1.0f", len_f);
