+++ src/ui_thing.cc
@@ -151,8 +151,8 @@ UI_ThingBox::UI_ThingBox(int X, int Y, int W, int H, const char *label) :
 	{
 		int dist = (i == 2 || i == 6) ? 32 : 35;
 
-		int x = ang_mx + dist * cos(i * 45 * M_PI / 180.0);
-		int y = ang_my - dist * sin(i * 45 * M_PI / 180.0);
+		int x = static_cast<int>(ang_mx + dist * cos(i * 45 * M_PI / 180.0));
+		int y = static_cast<int>(ang_my - dist * sin(i * 45 * M_PI / 180.0));
 
 		ang_buts[i] = new Fl_Button(x - 9, y - 9, 24, 24, 0);
 
@@ -813,9 +813,9 @@ void UI_ThingBox::UpdateField(int field)
 			const Thing *T = Things[obj];
 
 			// @@ FIXME show decimals in UDMF
-			pos_x->value(Int_TmpStr(T->x()));
-			pos_y->value(Int_TmpStr(T->y()));
-			pos_z->value(Int_TmpStr(T->h()));
+			pos_x->value(Int_TmpStr(static_cast<int>(T->x())));
+			pos_y->value(Int_TmpStr(static_cast<int>(T->y())));
+			pos_z->value(Int_TmpStr(static_cast<int>(T->h())));
 		}
 		else
 		{
